embedded video support

+++ FlashCardCreator/AWSS3UploadHelper.m
@@ -103,7 +103,7 @@ typedef NS_ENUM(NSUInteger, Type_ActionSheet) {
         
     } else {
         NSDictionary *downloadLinkageDict = [[NSUserDefaults standardUserDefaults] objectForKey:@"savedDownloadLinkage"];
-        NSString *linkage = [downloadLinkageDict objectForKey:[NSString stringWithFormat:@"%d",self.currentPack.packID]];
+        NSString *linkage = [downloadLinkageDict objectForKey:[NSString stringWithFormat:@"%ld",(long)self.currentPack.packID]];
         
         NSString *finalLinkage = [linkage stringByReplacingOccurrencesOfString:@"https://dl" withString:@"fcc://www"];
         finalLinkage = [finalLinkage stringByReplacingOccurrencesOfString:@"https://" withString:@"fcc://"];
