+++ FlashCardCreator/DropboxSharekitHelper.m
@@ -84,7 +84,7 @@
         
     } else {
         NSDictionary *downloadLinkageDict = [[NSUserDefaults standardUserDefaults] objectForKey:@"savedDownloadLinkage"];
-        NSString *linkage = [downloadLinkageDict objectForKey:[NSString stringWithFormat:@"%d",self.currentPack.packID]];
+        NSString *linkage = [downloadLinkageDict objectForKey:[NSString stringWithFormat:@"%ld",(long)self.currentPack.packID]];
         
         NSString *finalLinkage = [linkage stringByReplacingOccurrencesOfString:@"https://dl" withString:@"fcc://www"];
         finalLinkage = [finalLinkage stringByReplacingOccurrencesOfString:@"https://" withString:@"fcc://"];
